Purging of the Registry struct
