+++ vach/src/global/mod.rs
@@ -1,4 +1,4 @@
 // Globally available exports
 pub mod header;
-pub mod registry;
+pub mod reg_entry;
 pub mod types;
