+++ vach/src/loader/archive.rs
@@ -2,21 +2,23 @@ use anyhow;
 use super::resource::Resource;
 use crate::global::{
 	header::{Header, HeaderConfig},
-	registry::{Registry, RegistryEntry},
+	reg_entry::{RegistryEntry},
+	types::{FlagType},
 };
 use std::{
-	io::{BufReader, Cursor, Read, Seek, Write},
+	io::{self, BufReader, Cursor, Read, Seek, SeekFrom, Write},
 	str,
 };
-use ed25519_dalek as esdalek;
+use ed25519_dalek::{self as esdalek, Verifier};
+use lz4_flex as lz4;
 use hashbrown::HashMap;
 
 #[derive(Debug)]
 pub struct Archive<T> {
 	header: Header,
-	registry: Registry,
 	handle: T,
 	key: Option<esdalek::PublicKey>,
+	entries: HashMap<String, RegistryEntry>,
 }
 
 impl Archive<Cursor<Vec<u8>>> {
@@ -24,9 +26,9 @@ impl Archive<Cursor<Vec<u8>>> {
 	pub fn empty() -> Archive<Cursor<Vec<u8>>> {
 		Archive {
 			header: Header::default(),
-			registry: Registry::empty(),
 			handle: Cursor::new(Vec::new()),
 			key: None,
+			entries: HashMap::new()
 		}
 	}
 }
@@ -42,34 +44,79 @@ impl<T: Seek + Read> Archive<T> {
 	) -> anyhow::Result<Archive<impl Seek + Read>> {
 		let header = Header::from_handle(&mut handle)?;
 		Header::validate(&header, &config)?;
-		let registry = Registry::from_handle(&mut handle, &header, config.public_key.is_some())?;
+		// Generate and store Registry Entries
+		let mut entries = HashMap::new();
+		for _ in 0..header.capacity {
+			let (entry, id) = RegistryEntry::from_handle(&mut handle, config.public_key.is_some())?;
+			entries.insert(id, entry);
+		};
 
 		Ok(Archive {
 			header,
-			registry,
 			handle: BufReader::new(handle),
 			key: config.public_key,
+			entries,
 		})
 	}
 
 	// Query functions
 	#[inline]
 	pub fn fetch(&mut self, id: &str) -> anyhow::Result<Resource> {
-		self.registry.fetch(id, &mut self.handle, &self.key)
+		let mut buffer = Vec::new();
+		let (flags, content_version) = self.fetch_write(id, &mut buffer)?;
+		Ok(Resource{
+			content_version,
+			flags,
+			data: buffer
+		})
 	}
 	#[inline]
-	pub fn fetch_write(&mut self, id: &str, target: impl Write) -> anyhow::Result<()> {
-		self.registry
-			.fetch_write(id, &mut self.handle, &self.key, target)
+	pub fn fetch_write(&mut self, id: &str, mut target: impl Write) -> anyhow::Result<(FlagType, u8)> {
+		if let Some(entry) = self.fetch_entry(id) {
+			let handle = &mut self.handle;
+			handle.seek(SeekFrom::Start(entry.location))?;
+
+			let mut take = handle.take(entry.offset);
+			let mut buffer = Vec::new();
+			take.read_to_end(&mut buffer)?;
+
+			// The path is part of the signature, so you do not mess with the addressed data
+			buffer.extend(id.as_bytes());
+
+			// Validate signature
+			if let Some(pub_key) = &self.key {
+				if let Err(error) = pub_key.verify(&buffer, &entry.signature) {
+					anyhow::bail!(format!(
+						"({}): Invalid signature found for leaf with ID: {}",
+						error, id
+					))
+				};
+			};
+
+			// Decompress
+			if entry.flags.contains(FlagType::COMPRESSED) {
+				io::copy(
+					&mut lz4::frame::FrameDecoder::new(buffer.take(entry.offset)),
+					&mut target,
+				)?;
+			} else {
+				io::copy(&mut buffer.take(entry.offset), &mut target)?;
+			};
+
+			Ok((entry.flags, entry.content_version))
+		} else {
+			anyhow::bail!(format!("Resource not found: {}", id))
+		}
 	}
 	#[inline]
-	pub fn fetch_entry(&mut self, id: &str) -> Option<&RegistryEntry> {
-		self.registry.fetch_entry(id)
+	pub fn fetch_entry(&mut self, id: &str) -> Option<RegistryEntry> {
+		match self.entries.get(id) {
+			 Some(entry) => { Some(entry.clone()) },
+			 None => None,
+		}
 	}
 	#[inline]
-	pub fn entries(&self) -> &HashMap<String, RegistryEntry> {
-		&self.registry.entries
-	}
+	pub fn entries(&self) -> &HashMap<String, RegistryEntry> { &self.entries }
 }
 
 impl Default for Archive<Cursor<Vec<u8>>> {
